requirements.txt
PyYAML
fastapi
pycryptodome
aiosqlite
SQLAlchemy
passlib
uvicorn
pydantic
python-dotenv
jinja2
aiofiles
python-multipart